# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bill'Lib",
        "description": "",
        "contact": {
            "email": "nathalietaffot@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/users": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register",
                "description": "Register",
                "operationId": "store",
                "requestBody": {
                    "description": "Enregistrement d'un nouvel utilisateur",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "votre nom",
                                        "type": "string",
                                        "format": "string",
                                        "example": "laporte"
                                    },
                                    "avatar": {
                                        "description": "votre photo",
                                        "type": "string",
                                        "format": "string",
                                        "example": "https://kdjfkd.png"
                                    },
                                    "code_postal": {
                                        "description": "votre code postal",
                                        "type": "string",
                                        "format": "90931",
                                        "example": "384489"
                                    },
                                    "email": {
                                        "description": "votre email",
                                        "type": "string",
                                        "format": "string",
                                        "example": "examples@gmail.com"
                                    },
                                    "password": {
                                        "description": "votre password",
                                        "type": "string",
                                        "format": "string",
                                        "example": "sdms"
                                    },
                                    "code": {
                                        "description": "votre code",
                                        "type": "string",
                                        "format": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur bien Creer."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notification/identifiant/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Create identifier of User",
                "description": "Create identifier of User",
                "operationId": "createIdentifiant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "identifient bien créé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/uploadFile/{id}": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "upload avatar file",
                "description": "upload avatar file",
                "operationId": "uploadAvatar",
                "requestBody": {
                    "description": "Telechargement de la photo de profil utilisateur",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "description": "votre photos de profil",
                                        "type": "file",
                                        "format": "image",
                                        "example": "https://image.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "avatar_path": {
                                            "type": "string",
                                            "example": "users/avatar/ghRfjbiJHOvnMBaeerGTwCbYxV0WEnRuRPFod9N3.jpg"
                                        },
                                        "avatar_url": {
                                            "type": "string",
                                            "example": "http://tchallenger.test/users/avatar/ghRfjbiJHOvnMBaeerGTwCbYxV0WEnRuRPFod9N3.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [
                            []
                        ]
                    }
                ]
            }
        },
...
        "/api/check/email": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "verification",
                "description": "send email",
                "operationId": "checkEmail",
                "requestBody": {
                    "description": "Envoie du mail de verification a un nouvel utilisateur",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "votre email",
                                        "type": "string",
                                        "format": "string",
                                        "example": "examples@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "mail bien envoyer."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}